Move base.py for catalog to __init__
