+++ python/tests/catalog/test_hive.py
@@ -31,7 +31,7 @@ from hive_metastore.ttypes import (
 )
 from hive_metastore.ttypes import Table as HiveTable
 
-from pyiceberg.catalog.base import PropertiesUpdateSummary
+from pyiceberg.catalog import PropertiesUpdateSummary
 from pyiceberg.catalog.hive import HiveCatalog
 from pyiceberg.exceptions import (
     NamespaceAlreadyExistsError,
